+++ project/src/components/reviews-list/reviews-list.tsx
@@ -2,14 +2,15 @@ import {Reviews} from '../../types/review';
 import ReviewsItem from '../reviews-item/reviews-item';
 
 type ReviewsListProps = {
-  reviews?: Reviews;
+  reviews: Reviews;
 }
 
 function ReviewsList({reviews}: ReviewsListProps): JSX.Element {
+
   return (
-    <><h2 className="reviews__title">Reviews &middot; <span className="reviews__amount">{reviews?.length}</span></h2>
+    <><h2 className="reviews__title">Reviews &middot; <span className="reviews__amount">{reviews.length}</span></h2>
       <ul className="reviews__list">
-        {reviews?.map((review) => (
+        {reviews.map((review) => (
           <ReviewsItem key={review.id}
             review={review}
           />
